lab3/cloud_functions/temperature_processor/requirements.txt
functions-framework==3.*
google-cloud-firestore==2.*
google-cloud-pubsub==2.*